refactor ft_atio

+++ ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: oissa <oissa@student.42amman.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 11:27:34 by oissa             #+#    #+#             */
-/*   Updated: 2024/08/31 14:29:01 by oissa            ###   ########.fr       */
+/*   Updated: 2024/11/03 12:30:12 by oissa            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	res;
+	int		i;
+	int		sign;
+	long	res;
 
 	i = 0;
 	sign = 1;
